+++ apps/client/tsconfig.json
@@ -11,7 +11,7 @@
   "include": [
     "src",
     "vite.config.ts",
-    "../../packages/design-system/src/utils/use-click-outside.ts"
+    "../../packages/design-system/src/hooks/use-click-outside.ts"
   ],
   "exclude": ["node_modules"]
 }
